+++ back/src/bsffs/resolvers/mutations/__tests__/createBsff.integration.ts
@@ -14,6 +14,8 @@ import {
 import makeClient from "../../../../__tests__/testClient";
 import { fullBsff } from "../../../fragments";
 import * as sirenify from "../../../sirenify";
+import { createFicheIntervention } from "../../../__tests__/factories";
+import prisma from "../../../../prisma";
 
 const sirenifyMock = jest
   .spyOn(sirenify, "sirenifyBsffInput")
@@ -111,6 +113,39 @@ describe("Mutation.createBsff", () => {
     ]);
     // check input is sirenified
     expect(sirenifyMock).toHaveBeenCalledTimes(1);
+  });
+
+  it("should create a bsff with a fiche d'intervention", async () => {
+    const operateur = await userWithCompanyFactory(UserRole.ADMIN);
+    const detenteur = await userWithCompanyFactory(UserRole.ADMIN);
+    const transporter = await userWithCompanyFactory(UserRole.ADMIN);
+    const destination = await userWithCompanyFactory(UserRole.ADMIN);
+    const ficheIntervention = await createFicheIntervention({
+      operateur,
+      detenteur
+    });
+    const { mutate } = makeClient(operateur.user);
+    const { data, errors } = await mutate<
+      Pick<Mutation, "createBsff">,
+      MutationCreateBsffArgs
+    >(CREATE_BSFF, {
+      variables: {
+        input: {
+          ...createInput(operateur, transporter, destination),
+          ficheInterventions: [ficheIntervention.id]
+        }
+      }
+    });
+    expect(errors).toBeUndefined();
+    const bsff = await prisma.bsff.findUniqueOrThrow({
+      where: { id: data.createBsff.id },
+      include: { ficheInterventions: true }
+    });
+    expect(bsff.ficheInterventions.length).toEqual(1);
+    expect(bsff.ficheInterventions.map(fi => fi.id)).toEqual([
+      ficheIntervention.id
+    ]);
+    expect(bsff.detenteurCompanySirets).toEqual([detenteur.company.siret]);
   });
 
   it("should disallow unauthenticated user from creating a bsff", async () => {